5.1.7 Перенесёт логику отрисовки точек маршрута в презентер.
